Add API's for converting FHIR to Parquet and then extracting ASCVD parameters from Parquet (work in progress)
